fix usage, mainly integer conversoins

+++ cli/marlin-cvm/src/configs/global.rs
@@ -8,4 +8,4 @@ pub const DEFAULT_ATTESTATION_PORT: u16 = 1300;
 pub const MIN_DEPOSIT_AMOUNT: u64 = 1; // Minimum 0.000001 USDC
 pub const MIN_WITHDRAW_AMOUNT: u64 = 1; // Minimum 0.000001 USDC
 
-pub const EXTRA_DECIMALS: i64 = 6;
+pub const EXTRA_DECIMALS: u32 = 6;
